requirement.txt
fastapi
uvicorn[standard]
kubernetes
langgraph
google-generativeai
python-dotenv
httpx